- changed code to deal with image mtx as int values - organized test code more

+++ src/main/java/bunwarpj/BSplineModel.java
@@ -290,7 +290,7 @@ public class BSplineModel implements Runnable
 	 * @param isTarget enables the computation of the derivative or not
 	 */
 	public BSplineModel (
-			final double [][]img,
+			final int [][]img,
 			final boolean isTarget)
 	{
 		// Initialize thread
